Add interactivity and functionality test.
For each button in the calculator, a variable was
added. There are now event listeners that change
the calculator's buttons color on mouseover
and mouseout.

The equal sign was temporarily made to console log
the firstNumber variable to test the new added
functions: getFirstNumber, and getSecondNumber.

Both functions make the event listeners on the
number's buttons store the number in question
in diferent arrays: firstNumber and secondNumber.