Add basic operation functionality.
The event listeners of the numbers were modified
so that they store data inside firstNumber array
before an operator is selected, and inside
secondNUmber array after an operator is selected.

Additionally, an operate() function was  created
to give the "=" button event listener the
capability of triggering an
operation between a pair of numbers.

The clear button also now completely resets the
calculator.