Change if statements in number event listeners.
Now the event listeners of the number buttons
use  array.lenght of firstNumber
array to choose the array to store the data.

This makes it so that the state of the operator
variable is irrelevant to this decision, in hope
 of this allowing the implementation of multiple
operations in the calculator when more changes
to the code are made.