Add multiple operations functionality.
Now the calculator supports multiple calculations
in a serial order. Like: 1 + 1 + 2, shows 4.

For this, the event listeners of the operators
were changed so that they check if there is a
second number in existence, if it is, they
perform the operation, make firstNumber equal to
the value of the "answer" variable, reset the
second number and allow the user to input a new
second number.

It still does not show the answer of the first
operation because it is overridden by the operator
being displayed. This is soon to be fixed.