Fix first display issue.
The caculator now displays numbers correctly
when performing multiple operations without
using the "equals" button.

For this, new html divs and css classes where
added so that the display of the operators is
different from the display of the numbers.

Additionally, the change of the condition in the
if statements in the numbers event listeners
caused a bug in which the first number could
only store one value. To fix this, the event
listeners now check if a new variable "indicator"
is set to true or false. This variable is
manipulated by the operator event listeners.

The following problem persists:
The operator symbol is still being displayed after
the final answer is displayed.